requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
